+++ apps/dotcom/client/src/routes.tsx
@@ -58,7 +58,7 @@ export const router = createRoutesFromElements(
 			<Route path={`/${READ_ONLY_PREFIX}/:roomId`} lazy={() => import('./pages/public-readonly')} />
 		</Route>
 		{/* begin tla */}
-		<Route lazy={() => import('./tla/providers/TlaClerkProvider')}>
+		<Route lazy={() => import('./tla/providers/TlaRootProviders')}>
 			<Route path="/q/local" lazy={() => import('./tla/pages/local')} />
 			<Route lazy={() => import('./tla/providers/TlaAppLoggedInProvider')}>
 				{/* If not redirected, then local */}
