+++ apps/dotcom/client/src/tla/layouts/TlaSidebarLayout/TlaSidebarLayout.tsx
@@ -1,7 +1,6 @@
 import { ReactNode } from 'react'
-import { EditorContext, TldrawUiContextProvider, useValue } from 'tldraw'
+import { useValue } from 'tldraw'
 import { globalEditor } from '../../../utils/globalEditor'
-import { components } from '../../components/TlaEditor/TlaEditor'
 import { TlaSidebar } from '../../components/TlaSidebar/TlaSidebar'
 import { useApp } from '../../hooks/useAppState'
 import { usePreventAccidentalDrops } from '../../hooks/usePreventAccidentalDrops'
@@ -16,27 +15,16 @@ export function TlaSidebarLayout({ children }: { children: ReactNode; collapsabl
 		[app]
 	)
 	const currentEditor = useValue('editor', () => globalEditor.get(), [])
-	const MaybeEditorProvider = currentEditor ? EditorContext.Provider : FakeProvider
-
-	function FakeProvider({ children }: { children: ReactNode }) {
-		return children
-	}
-	const MaybeUiContextProvider = currentEditor ? TldrawUiContextProvider : FakeProvider
-
 	usePreventAccidentalDrops()
 
 	return (
-		<MaybeEditorProvider value={currentEditor}>
-			<MaybeUiContextProvider components={components}>
-				<div
-					className={styles.layout}
-					data-sidebar={isSidebarOpen}
-					data-sidebarmobile={isSidebarOpenMobile}
-				>
-					<TlaSidebar />
-					{children}
-				</div>
-			</MaybeUiContextProvider>
-		</MaybeEditorProvider>
+		<div
+			className={styles.layout}
+			data-sidebar={isSidebarOpen}
+			data-sidebarmobile={isSidebarOpenMobile}
+		>
+			{currentEditor && <TlaSidebar />}
+			{children}
+		</div>
 	)
 }
